+++ crates/ic-query/src/sns/report/source/model/token.rs
@@ -13,16 +13,16 @@ use crate::sns::report::{SnsTokenMetadataRow, SnsTokenStandardRow};
 ///
 
 #[derive(Clone, Debug, Eq, PartialEq)]
-pub(in crate::sns::report) struct MainnetSnsToken {
-    pub(in crate::sns::report) token_name: String,
-    pub(in crate::sns::report) token_symbol: String,
-    pub(in crate::sns::report) decimals: u8,
-    pub(in crate::sns::report) transfer_fee: String,
-    pub(in crate::sns::report) total_supply: String,
-    pub(in crate::sns::report) minting_account_owner: Option<String>,
-    pub(in crate::sns::report) minting_account_subaccount_hex: Option<String>,
-    pub(in crate::sns::report) ledger_index_canister_id: Option<String>,
-    pub(in crate::sns::report) ledger_index_error: Option<String>,
-    pub(in crate::sns::report) supported_standards: Vec<SnsTokenStandardRow>,
-    pub(in crate::sns::report) metadata: Vec<SnsTokenMetadataRow>,
+pub struct MainnetSnsToken {
+    pub token_name: String,
+    pub token_symbol: String,
+    pub decimals: u8,
+    pub transfer_fee: String,
+    pub total_supply: String,
+    pub minting_account_owner: Option<String>,
+    pub minting_account_subaccount_hex: Option<String>,
+    pub ledger_index_canister_id: Option<String>,
+    pub ledger_index_error: Option<String>,
+    pub supported_standards: Vec<SnsTokenStandardRow>,
+    pub metadata: Vec<SnsTokenMetadataRow>,
 }